based on feedback in the pull request I moved the newly created scroll helper file to the utils folder
